esm/fips: refactor disable into repo parent class
Add new RepoEntitlement.remove_apt_config which can now be called during
process_contract_deltas to clean out apt config prior to re-enable call.
